% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanratiosParam-class.R
\docType{class}
\name{meanratiosParam-class}
\alias{meanratiosParam-class}
\alias{MeanratiosParam-class,}
\alias{MeanRatiosParam-class}
\title{meanratiosParam-class}
\arguments{
\item{assayName}{Name of expression matrix in SingleCellExperiment assays 
(e.g. "counts").}

\item{singleCellExperiment}{Object of type SingleCellExperiment (see 
\code{?SingleCellExperiment}).}

\item{cellTypeVariable}{Name of cell type variable in SingleCellExperiment 
coldata.}
}
\description{
class definition for meanratiosParam, which uses 
DeconvoBuddies::get_mean_ratio2().
}
\details{
Main constructor for class \linkS4class{meanratiosParam}.
}
\examples{
exampleList <- getDeconvolutionExampleData()
singleCellExperimentExample <- randomSingleCellExperiment()
newParam <- meanratiosParam(singleCellExperiment=singleCellExperimentExample, cellTypeVariable="celltype", markersPerType=5)
markers <- typemarkers(newParam)

}
\references{
Louise Huuki-Meyers. DeconvoBuddies. (2023). R package. GitHub URL https://github.com/LieberInstitute/DeconvoBuddies.
}
\seealso{
\linkS4class{typemarkersParam}
}
