% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanratiosParam-class.R
\name{meanratiosParam}
\alias{meanratiosParam}
\title{Make new object of class meanratiosParam}
\usage{
meanratiosParam(
  singleCellExperiment,
  assayName = "counts",
  cellTypeVariable = "cellType",
  markersPerType = 20,
  returnInfo = FALSE
)
}
\arguments{
\item{singleCellExperiment}{Object of type SingleCellExperiment (see 
\code{?SingleCellExperiment}).}

\item{assayName}{Name of expression matrix in SingleCellExperiment assays 
(e.g. "counts").}

\item{cellTypeVariable}{Name of cell type variable in SingleCellExperiment 
coldata.}

\item{markersPerType}{Number of top markers to get per cell type.}

\item{returnInfo}{Whether to return metadata and original method outputs 
with predicted proportions.}
}
\value{
Object of class \linkS4class{meanratiosParam}
}
\description{
Main constructor for class \linkS4class{meanratiosParam}.
}
\details{
Main class for mapping arguments to the mean ratios method 
implemented as \code{DeconvoBuddies::get_mean_ratio2()}.
}
\examples{
exampleList <- getDeconvolutionExampleData()
singleCellExperimentExample <- randomSingleCellExperiment()
newParam <- meanratiosParam(singleCellExperiment=singleCellExperimentExample, cellTypeVariable="celltype", markersPerType=5)
markers <- typemarkers(newParam)
markers

}
\references{
Louise Huuki-Meyers. DeconvoBuddies. (2023). R package. GitHub URL https://github.com/LieberInstitute/DeconvoBuddies.
}
\seealso{
\linkS4class{typemarkersParam}
}
